Fixed translations when not using lang templates Call to translate failed because of missing 'cr' as first arg.
However, the original implementation lead to possible translation
leaks between reports. Indeed, the call to translate does not refer to
any res_id, so translate returned the first report string with the
given source, ignoring the report's ID. This lead to issues since the
same English wording can be translated differently depending on the
context.