Improvements on multi-record reports
Added a process_separately parameter to define whether we want to
handle multi-records by processing templates separately and merge them
with PDFtk or to process it once by passing a single "objects" list,
like it was possible in the original report_aeroo.

Changed ValidationError to UserError which is more appropriate.

Added need for PDFtk in README.